move stories and tests out of addon-docs templates
