+++ code/.storybook/main.ts
@@ -56,10 +56,6 @@ const config = defineMain({
       titlePrefix: 'highlight',
     },
     {
-      directory: '../addons/docs/src/blocks',
-      titlePrefix: 'addons/docs/blocks',
-    },
-    {
       directory: '../addons/a11y/src',
       titlePrefix: 'addons/accessibility',
     },
@@ -69,6 +65,10 @@ const config = defineMain({
     },
     {
       directory: '../addons/docs/template/stories',
+      titlePrefix: 'addons/docs',
+    },
+    {
+      directory: '../addons/docs/src',
       titlePrefix: 'addons/docs',
     },
     {
