+++ code/addons/docs/src/blocks/blocks/Unstyled.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { withMdxComponentOverride } from './withMdxComponentOverride';
+import { withMdxComponentOverride } from './with-mdx-component-override';
 
 const UnstyledImpl: React.FC<
   React.DetailedHTMLProps<React.HTMLAttributes<HTMLDivElement>, HTMLDivElement>
