+++ code/addons/docs/src/blocks/blocks/with-mdx-component-override.tsx
